Create MirrorData.ps1 for configure-vm-drive-mirroring
This is the script that runs on the schedule. It uses Robocopy with the /COPYALL switch to copy all file information (including NTFS security) and /MT to use multiple threads. The script excludes the ProgramData folder on both source and target. It includes try–catch logic and sets an exit code accordingly.